Training endpoint

+++ src/back-end/thread_classes/worker_consumer_thread.py
@@ -6,7 +6,7 @@ import threading
 import time
 from services.cancellation_handler import CancellationHandler
 from services.cancel_request import CancelRequest
-from services.nirdizati_wrapper import predict
+from services.nirdizati_wrapper import predict, train
 from services.queue_controller import subscribeToQueue, sendCancelRequest, sendTaskToQueue
 from services.task import Task
 import services.file_handler as fh
@@ -32,10 +32,11 @@ class WorkerConsumerThread(threading.Thread):
         if self.cancellations.hasCancel(received_task.taskID):
             self.cancellations.removeCancel(received_task.taskID)
 
-            # shutil.rmtree(received_task.predictors_path)
-            # os.remove(received_task.schema_path)
-            # os.remove(received_task.event_log_path)
-            fh.removePredictTaskFile(received_task.taskID)
+            # TRAINING/PREDICTION SPLIT
+            if received_task.predictors_path:
+                fh.removePredictTaskFile(received_task.taskID)
+            else:
+                fh.removeTrainingTaskFile(received_task.taskID)
 
             print(f"Task with ID: {received_task.taskID} present in cancel set.\n"
                   f"Removed files corresponding to the task.\n"
@@ -52,10 +53,17 @@ class WorkerConsumerThread(threading.Thread):
             sendTaskToQueue(received_task, "output")
 
             predictor_abs = os.path.join(os.getcwd(), received_task.predictors_path)
+            config_abs = os.path.join(os.getcwd(), received_task.config_path)
+            schema_abs = os.path.join(os.getcwd(), received_task.schema_path)
             eventlog_abs = os.path.join(os.getcwd(), received_task.event_log_path)
-            output_abs = os.path.join(os.getcwd(), fh.loadPredictRoot(received_task.taskID),received_task.taskID)
+            prediction_output_abs = os.path.join(os.getcwd(), fh.loadPredictRoot(received_task.taskID), received_task.taskID)
+            training_output_abs = os.path.join(os.getcwd(), fh.loadTrainingRoot(received_task.taskID), received_task.taskID)
             
-            p = mp.Process(target=predict, args=(predictor_abs, eventlog_abs, output_abs))
+            # TRAINING/PREDICTION SPLIT
+            if received_task.predictors_path:
+                p = mp.Process(target=predict, args=(predictor_abs, eventlog_abs, prediction_output_abs))
+            else:
+                p = mp.Process(target=train, args=(config_abs, schema_abs, eventlog_abs, training_output_abs))
 
             p.start()
 
@@ -69,10 +77,13 @@ class WorkerConsumerThread(threading.Thread):
                     while p.is_alive():
                         time.sleep(0.1)
                     received_task.setStatus(Task.Status.CANCELLED)
-                    # shutil.rmtree(received_task.predictors_path)
-                    # os.remove(received_task.schema_path)
-                    # os.remove(received_task.event_log_path)
-                    fh.removePredictTaskFile(received_task.taskID)
+
+                    # TRAINING/PREDICTION SPLIT
+                    if received_task.predictors_path:
+                        fh.removePredictTaskFile(received_task.taskID)
+                    else:
+                        fh.removePredictTrainingFile(received_task.taskID)
+                    
                     sendCancelRequest(CancelRequest(received_task.taskID, True), self.cancellations.corr_id)
                     channel.basic_ack(delivery_tag=method.delivery_tag)
                     print(f"Cancelled current task with ID: {received_task.taskID}.")
@@ -84,13 +95,18 @@ class WorkerConsumerThread(threading.Thread):
                     received_task.setStatus(Task.Status.COMPLETED)
                     print(f"Finished processing task: {received_task.taskID}")
                     sendTaskToQueue(received_task, "output")
-                    shutil.rmtree(received_task.predictors_path, onerror = lambda func, path, excinfo : print(excinfo))
+                    
                     try:
                         os.remove(received_task.schema_path)
                         os.remove(received_task.event_log_path)
+                        # TRAINING/PREDICTION SPLIT
+                        if received_task.predictors_path:
+                            shutil.rmtree(received_task.predictors_path, onerror = lambda func, path, excinfo : print(excinfo))
+                        else:
+                            fh.zipFile(received_task.taskID, keep_files = False)
                     except OSError as err:
                         print(err)
-                    # fh.removePredictTaskFile(received_task.taskID)
+
                     channel.basic_ack(delivery_tag=method.delivery_tag)
                     print("Waiting for a new task...")
                     return
